ci(metrics.yml): add step to install dependencies and start Docker service
Add a new step to install necessary dependencies such as Docker, jq,
curl, and git. Start and enable the Docker service to ensure it runs
properly. Add the current user to the Docker group to allow Docker
commands to be executed without sudo. This change ensures that the
workflow has all required tools and permissions to execute subsequent
steps successfully.